# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "bp-carnet",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/auth/signup": {
      "post": {
        "summary": "サインアップを受け付け、メールを送る。",
        "description": "応答は登録の有無に関わらず同じにし、違いはメールの中身だけにする (メールアドレスから利用の\n有無を探らせないため)。未登録・確認待ちなら確認リンク、確認済みならログインの案内を送る。\n確認リンクを直前に送っていれば、発行し直さず何も送らない。",
        "operationId": "signup",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "受け付けた (登録済みのメールアドレスでも同じ応答)"
          },
          "400": {
            "description": "リクエストボディが不正 (JSON として解釈できない等)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "Content-Type が application/json でない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "メールアドレスとして解釈できない、または必須フィールドが無い",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "同じ接続元からの要求が上限を超えた",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/signup/check": {
      "post": {
        "summary": "確認リンクがまだ使えるかを、使用済みにせずに確かめる (リンクを開いたときに使う)。",
        "operationId": "check_signup_link",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckLinkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "リンクは使える"
          },
          "400": {
            "description": "リンクが無効・期限切れ・使用済み、またはリクエストボディが不正",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "Content-Type が application/json でない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "必須フィールドが無い",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/signup/complete": {
      "post": {
        "summary": "確認リンクの先でパスワードを設定し、メールアドレスを確認済みにしてログインさせる。",
        "operationId": "complete_signup",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteSignupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "パスワードを設定し、ログインした"
          },
          "400": {
            "description": "リンクが無効・期限切れ・使用済み、またはリクエストボディが不正",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "description": "Content-Type が application/json でない",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "パスワードがポリシーを満たさない、または必須フィールドが無い",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "同じ接続元、または全体からの要求が上限を超えた",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CheckLinkRequest": {
        "type": "object",
        "description": "リンクの先を開いたときに、リンクがまだ使えるかを尋ねる要求。",
        "required": [
          "token"
        ],
        "properties": {
          "token": {
            "type": "string",
            "description": "メールのリンクに載っていたトークン。"
          }
        }
      },
...
      "CompleteSignupRequest": {
        "type": "object",
        "required": [
          "token",
          "password"
        ],
        "properties": {
          "password": {
            "type": "string"
          },
          "token": {
            "type": "string",
            "description": "確認メールのリンクに載っていたトークン。"
          }
        }
      },
...
      "ErrorBody": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "rows": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "$ref": "#/components/schemas/RowError"
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "description": "レスポンスボディの共通 envelope。`code` が機械可読な契約 (frontend はこちらで表示文言を引く)、\n`message` はローカライズしないデバッグ用の説明文。`rows` は `BatchValidation` のときだけ埋まる。",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        }
      },
...
      "RowError": {
        "type": "object",
        "description": "バッチ処理で 1 行の検証が失敗したことを表す。`index` はリクエスト配列内の 0 始まりの位置。",
        "required": [
          "index",
          "message"
        ],
        "properties": {
          "index": {
            "type": "integer",
            "minimum": 0
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "SignupRequest": {
        "type": "object",
        "required": [
          "email"
        ],
        "properties": {
          "email": {
            "type": "string",
            "example": "kyoko@example.com"
          }
        }
      },
...
    }
  }
}